glsl-uniform-out-of-bounds-2: silence warning
The buffer for the "name"-variable is 99 bytes large, which makes it
look like the name can be much larger than 9 bytes large. So let's just
make the array as large as the maximum string can become with a 98
character name, to silence the compiler.

Reviewed-by: Pierre-Eric Pelloux-Prayer <pierre-eric.pelloux-prayer@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/338>