util/getopt: drop incorrect license-violation warning
Whether or not it's a license-violation to link statically to LGPL code
depends on the license of the code that calls it. In our case, it's all
open source, so people can easily re-link the result, and this isn't a
legal issue.

So let's drop the scary warning, as it's incorrect for us.

Reviewed-by: Eric Anholt <eric@anholt.net>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/364>