ext_external_objects: new test to share a Vulkan vertex buffer
In this test a vertex buffer is allocated and filled with data by
Vulkan. Then OpenGL uses it to render a chess board image on screen. We
validate that we see the expected image (depends on the geometry) and
then we attempt to overwrite the buffer. Invalid operation error should
be generated and the contents should remain unchanged.

v2:
* code that checks if the final image is a chess board in a function had
  been moved in a function to be reused twice
* fixed comments style

v3:
* made the uninteresting_data array static so that it's automatically
  initialized to 0 and removed a loop that was filling it with data

Signed-off-by: Eleni Maria Stea <estea@igalia.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>