ext_external_objects: new test to reuse several times a Vulkan VBO
In this test a VBO is allocated and filled by Vulkan with vertices that
create quads that follow a checkerboard pattern. Then this VBO is used
by OpenGL for rendering and we check that the expected pattern had been
generated and then it is used by Vulkan in a renderpass that uses
similar shaders and we verify that the same checkerboard was displayed
again.

Contains some changes to the Vulkan framework and to the previous tests
because it modifies vk_draw to use VBOs when available.

v2:
* replaced memset with = {0} at initialization of dev_sz in vk.c
* removed leftover comment
* fixed comments style

v3:
* renamed dev_sz to offsets in vkCmdBindVertexBuffers

Signed-off-by: Eleni Maria Stea <estea@igalia.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>