ext_external_objects: require export feature flag only when required
Currently framework requires external memory feature for all images,
even when export did not happen. As example vk-image-display that
requires export for depth attachment even though it is not required.

This patch makes it possible for tests to set which buffers require
export and thus makes it possible for drivers that would not support
depth/stencil export to pass tests that do not require it.

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>
Reviewed-by: Eleni Maria Stea <estea@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/397>