write-after-clear: Test multi-layer clears
Demonstrates a bug in iris. Multi-layer clears on color surfaces are
currently skipped when the driver realizes that the first layer doesn't
need to be cleared. The driver should also be taking into account the
fact that other layers are involved in the clear operation.

Reviewed-by: Tapani Pälli <tapani.palli@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/398>