framework/replay: don't assert, rather raise
When we try to download a trace but no URL is provided and the file is
not already in the file system we would assert. Let's rather raise an
exception.

Likewise, when checking for the existence of dumped files in
compare_replay.

v2:
  - Replace assert with raise also in compare_replay.

Signed-off-by: Andres Gomez <agomez@igalia.com>
Reviewed-by: Dylan Baker <dylan@pnwbakers.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/353>