ext_external_objects/vk-pix-buf-update-errors: initialization fix
Used calloc instead of malloc to fill the data in piglit_display of
vk_pix_buf_update_errors.c. These data will only be used to see if
glBufferSubData will return an error and are of no importance.

v2: Also moved the call to glTexImage2D to take place after the
attempt to overwrite the buffer in order to check that it wasn't
overwritten.

v3: Dropped the cast in calloc

Signed-off-by: Eleni Maria Stea <estea@igalia.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>