nv_primitive_restart: Add tests for error conditions
All tests pass on Mesa.

v2: Don't call glGetError between glBegin and glEnd because it is not
allowed.

v3: Don't expect an error for glPrimitiveRestartIndex after a glBegin
that is put in a display list. The glBegin is not "executed," so it does
not affect the glPrimitiveRestartIndex call or the glGetError call.

Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/427>