arb_texture_view: Test NaN for some RGB32 clears
Before this patch, this test cleared a GL_RGB32F texture to a NaN
(0xffffffff), then checked for that exact bit pattern from the
texelFetch instruction. It's possible that HW may return a different NaN
encoding, so update the clear program to check for NaN instead.

Enables iris to pass this test on gen11+.

Closes: https://gitlab.freedesktop.org/mesa/piglit/-/issues/42
Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/412>