tests: optimize atomicCompSwap tests by exiting ASAP
When hw spawns a lot of threads, the contention on a single atomic
variable ("value") may be so high that the shader can take so much
time that it can be mistakenly recognized as a GPU hang.
This is what exactly happened when Intel Mesa 3D driver started
to use SIMD32 mode (which increases the number of concurrent
invocations) instead of SIMD16.

As there's no point in testing values bigger than the mask array can
hold, exit as soon as we are sure that the value will not be evaluated.

v2 (Caio):
* cast to matching types
* use >= instead of >

v3 (Marcin):
* check value inside of the compare & swap loop, still hitting
  the original bug (https://bugs.freedesktop.org/show_bug.cgi?id=106774)

v4 (Caio):
* move even more code to the compare & swap loop

v5 (Caio):
* improve comment

Closes mesa#3753.

Signed-off-by: Marcin Ślusarz <marcin.slusarz@intel.com>
Tested-by: Clayton Craft <clayton.a.craft@intel.com> (v1)
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/411>