glsl-1.50,tess: Test passing maximum count of varyings between stages
Tests:
 vs -> fs
 vs -> tcs
 vs -> tes
 vs -> gs
 gs -> fs
 tcs -> tes
 tcs -> tes (patches)
 tes -> gs
 tes -> fs

There are tests with similar intention in VK-GL-CTS:
 *.core.tessellation_shader.tessellation_shader_tessellation.max_in_out_attributes
 *.core.geometry_shader.limits.max_input_components
 *.core.geometry_shader.limits.max_output_components
 *.core.geometry_shader.limits.max_output_components_single_point
And in piglit:
 glsl-max-varyings tests vs -> fs
 max-input-components.shader_test test vs -> gs

However, they are a bit more complex than strictly necessary
to test the limits. Which, e.g. on freedreno, leads to a register
allocation failure with some of them.

Signed-off-by: Danylo Piliaiev <dpiliaiev@igalia.com>
Acked-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/430>