arb_fragment_program: Test KIL using result of SLT
Reproduces a bug caused by a bad optimization that would cause part of
the condition for the KIL instruction to be omitted.  This test uses a
similar pattern to a shader from Euro Truck Simulator 2 that is in my
shader-db.

I believe this is the underlying problem of
https://gitlab.freedesktop.org/mesa/mesa/-/issues/1346.  There will be
more information in a Mesa MR that is coming soon.

This test fails on versions of Mesa that include 68420d8322c ("nir:
Simplify min and max of b2f") and 9167324a86b ("nir/algebraic: Mark some
logic-joined comparison reductions as exact").

Acked-by: Eric Anholt <eric@anholt.net>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/454>