ext_external_objects: Another semaphores test
This test is related to an intel iris edge case reported by Mario
Kleiner in mesa MR!4337. When the rendering workload on a shared
resource is relatively low and glSignalSemaphoreEXT is followed by
glFlush there's an error generated on anv.

The test renders on the common resource using OpenGL and then
and calls glSignalSemaphoreEXT and draws with Vulkan. The error
mentioned in MR!4337 is generated without the provided patch and
disappears with it.

Signed-off-by: Eleni Maria Stea <elene.mst@gmail.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>