ext_external_objects: New semaphores test
The test helps debugging an edge case we detected on intel iris fences
backend. When a submission workload is low and glFlush follows glSignal
VK_ERROR is generated on anv. There's a patch by Mario Kleiner on
MR!4337 that fixes the problem.

The test submits a low workload Vulkan queue and then renders in the
common resource using OpenGL. An assertion failure when stale objects
are cleared on iris without the patch (see mesa MR!4337) and is fixed
with it.

Signed-off-by: Eleni Maria Stea <elene.mst@gmail.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>