ext_external_objects: fix validation error on vertex binding descriptions
This fixes validation error VUID-vkCmdDraw-None-04007 thrown when
running tests with validation layer.

Error is thrown because pipeline is setup with a binding description
but vkCmdBindVertexBuffers is not called. Fix is to skip setting up
binding description when vbo is not being used by the test.

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>
Reviewed-by: Eleni Maria Stea <estea@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/464>