framework/replay: do not raise in compare if no images were dumped
Previously we were trusting that the backends will always return a
failure in case replaying a trace was not succesfull.

Now, we manage also this scenario without exceptions because
eglretrace won't fail if the parameters passed are malformed and we
really want to deal with it gracefully in any case.

Signed-off-by: Andres Gomez <agomez@igalia.com>
Reviewed-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>
Acked-by: Martin Peres <martin.peres@mupuf.org>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/466>