arb_gpu_shader_int64: test masking after addition
There is an optimization in Mesa that tries to reassociate these
operations to improve the chances of CSE. However, the base mask used
is 0xffffffff.  This is not correct (but should still work) for 16-bit
and 8-bit values, but it means the high 32-bits of 64-bit values won't
get chopped off.

Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/469>