util/dma_buf: Prefer gbm over libdrm-intel
The Intel OpenGL drivers and gbm are both part of the Mesa project.
Therefore, assuming the OpenGL driver supports the graphics device,
then gbm should support it as well.

The libdrm installed on the system may be older and unable to
recognize the graphics device. This can lead to the dmabuf tests
failing simply due to an outdated libdrm.

For this reason, it is preferable to rely on gbm if it available.

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/472>