spec/glsl-1.20: Use min and max with NaN
In IEEE-754 and OpenCL, min and max with a single NaN parameter are
required to return the other paramter.  GLSL does not have this
requirement, but numerous applications depend on this behavior.  These
tests try to exercise this behavior in a way that should be compatible
GLSL implementations that cannot produce NaN or Inf.

Reproduces the failure in mesa#4254.

Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/475>