ext_external_objects: All supported image usage flags are set
According to issue 13 of EXT_external_objects extension all supported
usage flags for an image should be set from Vulkan if the image is going
to be used from OpenGL. Change the function of Vulkan framework that
sets the usage flags to check which flags are supported for the format
and the tiling and use them at image creation automatically. Modified
the tests so that they don't expect the flag to be set by the user.

Signed-off-by: Eleni Maria Stea <elene.mst@gmail.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>