ext_image_dma_buf_import: Bump libdrm requirement to 2.4.98
2.4.98 was released two years ago, so this seems safe.  I discussed this
with folks on IRC #dri-devel, and everyone was okay with the change.
Ajax said that it would mean folks building piglit on RHEL would need to
build their own libdrm, but the shouldn't be a problem.  The Intel CI
always builds the latest libdrm, so this shouldn't cause any problems
there.

Reviewed-by: Tapani Pälli <tapani.palli@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/501>