glsl-1.30: Test range analysis of saturate in the presence of NaN
This test is a bit more tricky than the `min` and `max` tests.  The
biggest problem, as noted in comments in the test case, is that there
are multiple bugs present in Mesa's handling of saturate.  Rhys noticed
the issue in range analysis.  This is essentially the same as the issue
for `min` and `max`.

Implementing this test discovered a problem in an algebraic optimization
that strips saturate from some comparisons with zero.

Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/463>