CI: Update libdrm beyond what is in Debian
This tries to be a little bit clever.  Another commit will bump the
required version to 2.4.98.  If the installed version is already greater
than 2.4.98, build-debian.sh will use the system version.

One bummer... the libdrm version is in both build-debian.sh and
debian-install.sh.  It would be nice to have the version in a single
place so that both files wouldn't need to be kept in lock-step.  Alas.

v2: Add missing '\' after unzip in debian-install.sh.  Bump DEBIAN_TAG.
Use curl instead of wget, and invoke it from debian-install.sh so that
the results can be cached.  All suggested by Jose.  Fix a typo in the
'pkg-config --max-version' line in build-debian.sh.

Acked-by: Jose Fonseca <jfonseca@vmware.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/501>